[*] вынем CmdBuilder в отдельный файл
