+++ sources/libs/translation_module/include/translation_module/cmd_translator.h
@@ -5,40 +5,12 @@
 #include <json/json.h>
 #include "aliases.h"
 #include "arg_builders.h"
+#include "cmd_builder.h"
 #include "cmd_info.h"
 #include "translation_module.h"
 #include "fillers.h"
 
 namespace trm {
-
-class CmdBuilder {
-public:
-    CmdBuilder(Json::Value &&cmdSample,
-               std::vector<std::shared_ptr<ArgBuilder>> &&argBuilders) :
-        cmdSample(std::move(cmdSample)),
-        argBuilders(std::move(argBuilders))
-    {
-    }
-
-    Json::Value createCmd()
-    {
-        Json::Value result(cmdSample);
-
-        if (!argBuilders.empty()) {
-            Json::Value &args = result["args"] = Json::Value(Json::arrayValue);
-
-            for (auto &argBuilder : argBuilders) {
-                args.append(argBuilder->createArg());
-            }
-        }
-
-        return result;
-    }
-
-private:
-    Json::Value cmdSample;
-    std::vector<std::shared_ptr<ArgBuilder>> argBuilders;
-};
 
 enum class ArgType {
     StringPattern,
